+++ src/test/java/domain/player/BattingMoneyTest.java
@@ -1,4 +1,4 @@
-package domain.game;
+package domain.player;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
